+++ projects/fusion-ui/components/dropdown-dual-multi-select/v2/components/dropdown-dual-multi-select-body/dropdown-dual-multi-select-body.component.ts
@@ -45,6 +45,7 @@ export class DropdownDualMultiSelectBodyComponent implements OnInit, OnDestroy, 
 
     @Input() totalItems: number;
     @Input() isPendingItems: boolean;
+    @Input() hasSelectAll: boolean = true;
 
     @Input() set items(data: DropdownOption[]) {
         this.options$.next(data || []);