+++ projects/fusion-ui/components/floating-toolbar/v3/floating-toolbar.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FloatingToolbarComponent } from './floating-toolbar.component';
+
+describe('FloatingToolbarComponent', () => {
+  let component: FloatingToolbarComponent;
+  let fixture: ComponentFixture<FloatingToolbarComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ FloatingToolbarComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(FloatingToolbarComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});