+++ projects/fusion-ui/components/input-inline/common/base/inline-input-type.enum.ts
@@ -1,5 +1,6 @@
 export enum InlineInputType {
     Text,
     Number,
-    Currency
+    Currency,
+    Percent
 }