+++ projects/fusion-ui/components/input-inline/v3/input-inline.component.stories.ts
@@ -934,3 +934,81 @@ export class FusionStoryWrapperComponent {
         }
     }
 };
+
+export const ValuePercent = InputTemplate.bind({});
+ValuePercent.args = {
+    formControl: new FormControl(50),
+    type: InlineInputType.Percent
+};
+ValuePercent.parameters = {
+    docs: {
+        source: {
+            language: 'typescript',
+            code: dedent`
+import { Component, ViewChild } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { IconModule } from '@ironsource/fusion-ui/components/icon/v1';
+import {
+  InlineInputType,
+  InputInlineComponent,
+} from '@ironsource/fusion-ui/components/input-inline';
+
+@Component({
+  selector: 'fusion-story-wrapper',
+  template: \`
+    <div style="width: 155px;">
+<fusion-input-inline #inputInline
+    [type]="type"
+    [formControl]="formControl"
+    [loading]="loading"
+    [readOnly]="readOnly"
+    [error]="error"
+    (onSave)="onSave($event)"
+    (onCancel)="onCancel()"
+></fusion-input-inline>
+</div>
+  \`,
+  standalone: true,
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    IconModule,
+    InputInlineComponent,
+  ],
+})
+export class FusionStoryWrapperComponent {
+  @ViewChild('inputInline') inputInline: InputInlineComponent;
+
+  type = InlineInputType.Percent;
+  formControl = new FormControl(50);
+
+  loading = false;
+  readOnly = false;
+  error = '';
+
+  onSave($event) {
+    console.log('onSave: ', $event);
+    this.loading = true;
+    setTimeout(() => {
+      // here you can apply validations
+
+      this.loading = false;
+      // update value
+      this.formControl.setValue($event.newValue);
+      // exit from edit mode
+      this.inputInline.isEditMode$.next(false);
+    }, 1000);
+  }
+
+  onCancel() {
+    console.log('onCancel');
+  }
+}
+`,
+            format: true,
+            type: 'code'
+        }
+    }
+};