+++ projects/fusion-ui/components/tags-input/v3/tags-input.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TagsInputComponent } from './tags-input.component';
+
+describe('TagsInputComponent', () => {
+  let component: TagsInputComponent;
+  let fixture: ComponentFixture<TagsInputComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ TagsInputComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(TagsInputComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});